requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.4
openai==0.20.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.61
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
uuid==1.30
whitenoise==6.2.0
